Added spawn logic
when player moves it emits a signal that is then received by the world, the world checks everytime they move if a random number is equal to the enemy_spawn_chance in the current_area_resource. if yes they need to battle that enemy (needs to be implemented). mercy_rule allows the player to move freely without spawning enemies. for safe areas change the go into battle function to only happen if enemy exists OR see if making spawn chance a negative value will prevent it from happening OR have a boolean in everything_in_area resource that checks if the area is safe or not